# Conflicts:
#	src/main/java/teamproject/gunha/controller/MainController.java
#	src/main/resources/templates/header/footer.html
#	src/main/resources/templates/header/header.html

+++ src/main/java/teamproject/gunha/service/ProfileServiceImpl.java
@@ -1,5 +1,7 @@
 package teamproject.gunha.service;
 
+import java.util.Map;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -8,10 +10,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.extern.slf4j.Slf4j;
+import teamproject.gunha.mapper.OrderMapper;
 import teamproject.gunha.mapper.ProfileMapper;
 import teamproject.gunha.mapper.UserMapper;
 import teamproject.gunha.security.config.auth.NetflixUserDetails;
 import teamproject.gunha.vo.ProfileVO;
+import teamproject.gunha.vo.UserVO;
 
 @Service
 @Slf4j
@@ -23,13 +27,21 @@ public class ProfileServiceImpl implements ProfileService {
   @Autowired
   private UserMapper userMapper;
 
+  @Autowired
+  private OrderMapper orderMapper;
+
+
   @Override
   public boolean createProfile(ProfileVO profileVO) {
     log.info(profileVO + "");
-    int resultRowLine = profileMapper.insertProfile(profileVO);
-    if (resultRowLine >= 1) {
-      refreshUser();
-      return true;
+    int count = profileMapper.getNumberOfUserProfile(profileVO);
+    if (count < 5) {
+      int resultRowLine = profileMapper.insertProfile(profileVO);
+      if (resultRowLine >= 1) {
+        refreshUser();
+        return true;
+      }
+
     }
     return false;
   }
@@ -49,20 +61,32 @@ public class ProfileServiceImpl implements ProfileService {
   @Override
   @Transactional
   public boolean removeProfile(ProfileVO profileVO) {
-    int resultRowLine = profileMapper.deleteProfile(profileVO);
-    if (resultRowLine >= 1) {
-      refreshUser();
-      return true;
+    int count = profileMapper.getNumberOfUserProfile(profileVO);
+    if (count > 1) {
+      int resultRowLine = profileMapper.deleteProfile(profileVO);
+      if (resultRowLine >= 1) {
+        refreshUser();
+        return true;
+      }
     }
     return false;
   }
 
+
+
+
   private void refreshUser() {
-    NetflixUserDetails netflixUserDetails = (NetflixUserDetails) SecurityContextHolder.getContext().getAuthentication()
+    NetflixUserDetails prevUserDetails = (NetflixUserDetails) SecurityContextHolder.getContext().getAuthentication()
         .getPrincipal();
-    netflixUserDetails.setUserVO(userMapper.selectUserId(netflixUserDetails.getUsername()));
+
+    UserVO userVO = userMapper.selectUserId(prevUserDetails.getUsername());
+    userVO.setLastOrder(orderMapper.selectUserLastOrder(userVO.getUserId()));
+    userVO.setSecondLastOrder(orderMapper.selectUserSecondLastOrder(userVO.getUserId()));
+    NetflixUserDetails netflixUserDetails = new NetflixUserDetails(userVO);
+    log.info(netflixUserDetails + "");
     Authentication authentication = new UsernamePasswordAuthenticationToken(netflixUserDetails,
         netflixUserDetails.getPassword(), netflixUserDetails.getAuthorities());
     SecurityContextHolder.getContext().setAuthentication(authentication);
   }
+
 }
